[feat/#5] 네비게이션바 생성
